refactor(1834): factorize ModaleListeAcces components
